rename to `register`, `estimate-registration`, and `optimize-registration`
